+++ Package.swift
@@ -159,7 +159,6 @@ var streamChatSourcesExcluded: [String] { [
     "Models/Attachments/ChatMessageAttachment_Tests.swift",
     "Models/Attachments/ImageAttachmentPayload_Tests.swift",
     "Models/ChatMessage_Tests.swift",
-    "Workers/Background/NewChannelQueryUpdater_Tests.swift",
     "Workers/Background/AttachmentUploader_Tests.swift",
     "Workers/Background/DatabaseCleanupUpdater_Mock.swift",
     "Workers/Background/NewUserQueryUpdater_Tests.swift",