+++ Sources/StreamChat/ChatClient_Tests.swift
@@ -45,7 +45,6 @@ class ChatClient_Tests: StressTestCase {
     
     var workerBuilders: [WorkerBuilder] = [
         MessageSender.init,
-        NewChannelQueryUpdater.init,
         NewUserQueryUpdater.init
     ]
     
@@ -597,7 +596,6 @@ class ChatClient_Tests: StressTestCase {
         
         // Check all the mandatory background workers are initialized
         XCTAssert(client.backgroundWorkers.contains { $0 is MessageSender })
-        XCTAssert(client.backgroundWorkers.contains { $0 is NewChannelQueryUpdater })
         XCTAssert(client.backgroundWorkers.contains { $0 is NewUserQueryUpdater })
         XCTAssert(client.backgroundWorkers.contains { $0 is MessageEditor })
         XCTAssert(client.backgroundWorkers.contains { $0 is ConnectionRecoveryUpdater })